+++ packages/common/src/api/tan-query/coins/useArtistCoins.ts
@@ -1,7 +1,8 @@
-import { encodeHashId } from '@audius/sdk'
+import { Id } from '@audius/sdk'
 import { useQuery } from '@tanstack/react-query'
 
 import { ID } from '~/models'
+import { removeNullable } from '~/utils/typeUtils'
 
 import { QUERY_KEYS } from '../queryKeys'
 import { useQueryContext } from '../utils/QueryContext'
@@ -23,11 +24,8 @@ export const useArtistCoins = (params: UseArtistCoinsParams = {}) => {
 
       // Encode owner_id params to match API expectations
       const encodedOwnerIds = params.owner_id
-        ?.map((id) => {
-          const encodedId = encodeHashId(id)
-          return encodedId
-        })
-        .filter((id): id is string => Boolean(id))
+        ?.map((id) => Id.parse(id))
+        .filter(removeNullable)
 
       const response = await sdk.coins.getCoins({
         mint: params.mint,
@@ -36,7 +34,7 @@ export const useArtistCoins = (params: UseArtistCoinsParams = {}) => {
         offset: params.offset
       })
 
-      return response.data
+      return response?.data
     }
   })
 }
