+++ packages/harmony/src/components/artwork/Artwork.tsx
@@ -54,197 +54,134 @@ export const Artwork = (props: ArtworkProps) => {
     }
   }, [src])
 
-  const artworkContent = (
-    <Box {...other}>
-      <Box
-        borderRadius={hex ? undefined : borderRadius}
-        border={hex ? undefined : 'default'}
-        shadow={shadow}
-        css={{ borderWidth: hex ? undefined : borderWidth }}
-      >
-        {isLoading ? (
-          <Skeleton
-            borderRadius={hex ? undefined : borderRadius}
-            h='100%'
-            w='100%'
-            css={{
-              zIndex: 2,
-              position: 'absolute',
-              opacity: isLoading ? 1 : 0,
-              transition: `opacity ${motion.calm}`,
-              clipPath: hex ? `url(#${roundedHexClipPath})` : undefined
-            }}
-          />
-        ) : null}
-        <Box
+  // Helper function to render the common artwork elements
+  const renderArtworkElements = (useHexClip: boolean) => (
+    <>
+      {isLoading ? (
+        <Skeleton
+          borderRadius={useHexClip ? borderRadius : borderRadius}
+          h='100%'
           w='100%'
-          pt='100%'
-          borderRadius={hex ? undefined : borderRadius}
           css={{
-            backgroundColor:
-              !src && children ? color.neutral.n400 : color.background.surface2,
-            clipPath: hex ? `url(#${roundedHexClipPath})` : undefined
+            zIndex: 2,
+            position: 'absolute',
+            opacity: isLoading ? 1 : 0,
+            transition: `opacity ${motion.calm}`,
+            clipPath: useHexClip ? `url(#${roundedHexClipPath})` : undefined
           }}
         />
-        {src ? (
-          <Box
-            as='img'
-            ref={imgRef}
-            borderRadius={hex ? undefined : borderRadius}
-            h='100%'
-            w='100%'
-            onLoad={() => {
-              setIsLoadingState(false)
-            }}
-            onError={(event) => {
-              setIsLoadingState(false)
-              onError?.(event)
-            }}
-            // @ts-ignore
-            src={src}
-            data-testid={testId}
-            draggable={false}
-            css={{
-              position: 'absolute',
-              top: 0,
-              left: 0,
-              objectFit: 'cover',
-              opacity: isLoading ? 0 : 1,
-              transition: `opacity ${motion.calm}`,
-              clipPath: hex ? `url(#${roundedHexClipPath})` : undefined
-            }}
-          />
-        ) : null}
-        {children ? (
-          <Flex
-            h='100%'
-            w='100%'
-            alignItems='center'
-            justifyContent='center'
-            css={{
-              position: 'absolute',
-              top: 0,
-              left: 0,
-              clipPath: hex ? `url(#${roundedHexClipPath})` : undefined
-            }}
-          >
-            {children}
-          </Flex>
-        ) : null}
-      </Box>
-    </Box>
+      ) : null}
+      <Box
+        w='100%'
+        pt='100%'
+        borderRadius={useHexClip ? borderRadius : borderRadius}
+        css={{
+          backgroundColor:
+            !src && children ? color.neutral.n400 : color.background.surface2,
+          clipPath: useHexClip ? `url(#${roundedHexClipPath})` : undefined
+        }}
+      />
+      {src ? (
+        <Box
+          as='img'
+          ref={imgRef}
+          borderRadius={useHexClip ? borderRadius : borderRadius}
+          h='100%'
+          w='100%'
+          onLoad={() => {
+            setIsLoadingState(false)
+          }}
+          onError={(event) => {
+            setIsLoadingState(false)
+            onError?.(event)
+          }}
+          // @ts-ignore
+          src={src}
+          data-testid={testId}
+          draggable={false}
+          css={{
+            position: 'absolute',
+            top: 0,
+            left: 0,
+            objectFit: 'cover',
+            opacity: isLoading ? 0 : 1,
+            transition: `opacity ${motion.calm}`,
+            clipPath: useHexClip ? `url(#${roundedHexClipPath})` : undefined
+          }}
+        />
+      ) : null}
+      {children ? (
+        <Flex
+          h='100%'
+          w='100%'
+          alignItems='center'
+          justifyContent='center'
+          css={{
+            position: 'absolute',
+            top: 0,
+            left: 0,
+            clipPath: useHexClip ? `url(#${roundedHexClipPath})` : undefined
+          }}
+        >
+          {children}
+        </Flex>
+      ) : null}
+    </>
   )
 
-  // If not hexagonal, return the artwork as-is
-  if (!hex) {
-    return artworkContent
+  // For hexagonal artwork, wrap with border and SVG overlay
+  if (hex) {
+    return (
+      <Box {...other} css={{ position: 'relative' }}>
+        <Box
+          borderRadius={borderRadius}
+          border='default'
+          shadow={shadow}
+          css={{ borderWidth }}
+        >
+          {renderArtworkElements(true)}
+        </Box>
+        {/* Hexagonal border overlay */}
+        <svg
+          width='100%'
+          height='100%'
+          viewBox='0 0 1 1'
+          css={{
+            position: 'absolute',
+            top: 0,
+            left: 0,
+            pointerEvents: 'none'
+          }}
+        >
+          <g
+            transform={`translate(${(1 - HEXAGON_BORDER_INSET_SCALE) / 2}, ${
+              (1 - HEXAGON_BORDER_INSET_SCALE) / 2
+            }) scale(${HEXAGON_BORDER_INSET_SCALE})`}
+          >
+            <path
+              d='M0.966 0.378C0.93 0.301 0.887 0.228 0.839 0.158L0.824 0.136C0.805 0.108 0.779 0.085 0.75 0.068C0.721 0.051 0.688 0.041 0.655 0.039L0.627 0.036C0.543 0.03 0.457 0.03 0.373 0.036L0.346 0.039C0.312 0.041 0.279 0.051 0.25 0.068C0.221 0.085 0.196 0.108 0.177 0.136L0.161 0.158C0.113 0.228 0.07 0.302 0.034 0.378L0.022 0.403C0.008 0.433 0 0.466 0 0.5C0 0.534 0.008 0.567 0.022 0.597L0.034 0.622C0.07 0.698 0.113 0.772 0.161 0.842L0.177 0.864C0.196 0.892 0.221 0.915 0.25 0.932C0.279 0.949 0.312 0.959 0.346 0.961L0.373 0.964C0.457 0.97 0.543 0.97 0.627 0.964L0.655 0.961C0.688 0.959 0.721 0.949 0.75 0.932C0.779 0.915 0.805 0.892 0.824 0.864L0.839 0.842C0.887 0.772 0.93 0.698 0.966 0.622L0.978 0.597C0.992 0.567 1 0.534 1 0.5C1 0.466 0.992 0.433 0.978 0.403L0.966 0.378Z'
+              fill='none'
+              opacity={0.3}
+              stroke={color.neutral.n950}
+              strokeWidth='0.005'
+            />
+          </g>
+        </svg>
+      </Box>
+    )
   }
 
-  // For hexagonal artwork, wrap with border
+  // For regular artwork
   return (
-    <Box {...other} css={{ position: 'relative' }}>
+    <Box {...other}>
       <Box
         borderRadius={borderRadius}
         border='default'
         shadow={shadow}
         css={{ borderWidth }}
       >
-        {isLoading ? (
-          <Skeleton
-            borderRadius={borderRadius}
-            h='100%'
-            w='100%'
-            css={{
-              zIndex: 2,
-              position: 'absolute',
-              opacity: isLoading ? 1 : 0,
-              transition: `opacity ${motion.calm}`,
-              clipPath: `url(#${roundedHexClipPath})`
-            }}
-          />
-        ) : null}
-        <Box
-          w='100%'
-          pt='100%'
-          borderRadius={borderRadius}
-          css={{
-            backgroundColor:
-              !src && children ? color.neutral.n400 : color.background.surface2,
-            clipPath: `url(#${roundedHexClipPath})`
-          }}
-        />
-        {src ? (
-          <Box
-            as='img'
-            ref={imgRef}
-            borderRadius={borderRadius}
-            h='100%'
-            w='100%'
-            onLoad={() => {
-              setIsLoadingState(false)
-            }}
-            onError={(event) => {
-              setIsLoadingState(false)
-              onError?.(event)
-            }}
-            // @ts-ignore
-            src={src}
-            data-testid={testId}
-            draggable={false}
-            css={{
-              position: 'absolute',
-              top: 0,
-              left: 0,
-              objectFit: 'cover',
-              opacity: isLoading ? 0 : 1,
-              transition: `opacity ${motion.calm}`,
-              clipPath: `url(#${roundedHexClipPath})`
-            }}
-          />
-        ) : null}
-        {children ? (
-          <Flex
-            h='100%'
-            w='100%'
-            alignItems='center'
-            justifyContent='center'
-            css={{
-              position: 'absolute',
-              top: 0,
-              left: 0,
-              clipPath: `url(#${roundedHexClipPath})`
-            }}
-          >
-            {children}
-          </Flex>
-        ) : null}
+        {renderArtworkElements(false)}
       </Box>
-      {/* Hexagonal border overlay */}
-      <svg
-        width='100%'
-        height='100%'
-        viewBox='0 0 1 1'
-        css={{
-          position: 'absolute',
-          top: 0,
-          left: 0,
-          pointerEvents: 'none'
-        }}
-      >
-        <g
-          transform={`translate(${(1 - HEXAGON_BORDER_INSET_SCALE) / 2}, ${
-            (1 - HEXAGON_BORDER_INSET_SCALE) / 2
-          }) scale(${HEXAGON_BORDER_INSET_SCALE})`}
-        >
-          <path
-            d='M0.966 0.378C0.93 0.301 0.887 0.228 0.839 0.158L0.824 0.136C0.805 0.108 0.779 0.085 0.75 0.068C0.721 0.051 0.688 0.041 0.655 0.039L0.627 0.036C0.543 0.03 0.457 0.03 0.373 0.036L0.346 0.039C0.312 0.041 0.279 0.051 0.25 0.068C0.221 0.085 0.196 0.108 0.177 0.136L0.161 0.158C0.113 0.228 0.07 0.302 0.034 0.378L0.022 0.403C0.008 0.433 0 0.466 0 0.5C0 0.534 0.008 0.567 0.022 0.597L0.034 0.622C0.07 0.698 0.113 0.772 0.161 0.842L0.177 0.864C0.196 0.892 0.221 0.915 0.25 0.932C0.279 0.949 0.312 0.959 0.346 0.961L0.373 0.964C0.457 0.97 0.543 0.97 0.627 0.964L0.655 0.961C0.688 0.959 0.721 0.949 0.75 0.932C0.779 0.915 0.805 0.892 0.824 0.864L0.839 0.842C0.887 0.772 0.93 0.698 0.966 0.622L0.978 0.597C0.992 0.567 1 0.534 1 0.5C1 0.466 0.992 0.433 0.978 0.403L0.966 0.378Z'
-            fill='none'
-            opacity={0.3}
-            stroke={color.neutral.n950}
-            strokeWidth='0.005'
-          />
-        </g>
-      </svg>
     </Box>
   )
 }
